flask_controller/requirements.txt
Flask
Gunicorn
Torch
bert-extractive-summarizer
sentence-transformers
